+++ src/frontend/src/lib/components/loaders/LoaderSwapTokens.svelte
