+++ src/frontend/src/lib/components/loaders/Loaders.svelte
@@ -12,7 +12,7 @@
 	import LoaderContacts from '$lib/components/loaders/LoaderContacts.svelte';
 	import LoaderHarvest from '$lib/components/loaders/LoaderHarvest.svelte';
 	import LoaderMetamask from '$lib/components/loaders/LoaderMetamask.svelte';
-	import LoaderSwapSupportedTokens from '$lib/components/loaders/LoaderSwapSupportedTokens.svelte';
+	import LoaderSwapTokens from '$lib/components/loaders/LoaderSwapTokens.svelte';
 	import LoaderTokens from '$lib/components/loaders/LoaderTokens.svelte';
 	import LoaderUserProfile from '$lib/components/loaders/LoaderUserProfile.svelte';
 	import LoaderWallets from '$lib/components/loaders/LoaderWallets.svelte';
@@ -56,7 +56,7 @@
 
 		<LoaderHarvest />
 
-		<LoaderSwapSupportedTokens />
+		<LoaderSwapTokens />
 
 		{@render children()}
 	</Loader>
